feat(packages/sui-studio): fix styles
