+++ packages/sui-studio/src/components/navigation/index.js
@@ -1,5 +1,3 @@
-import {useState} from 'react'
-
 import PropTypes from 'prop-types'
 
 import {Link} from '@s-ui/react-router'
@@ -8,17 +6,7 @@ import {getComponentsList} from '../utils.js'
 
 const componentsList = getComponentsList()
 
-export default function Navigation({handleClick}) {
-  const [search, setSearch] = useState('')
-
-  const handleChange = e => {
-    setSearch(e.target.value)
-  }
-
-  const handleFocus = e => {
-    e.target.select()
-  }
-
+export default function Navigation({handleClick, search}) {
   const filterComponentsFromSearch = ({search}) => {
     return componentsList.filter(({category, component}) => {
       const lowerCasedSearch = search.toLowerCase()
@@ -65,15 +53,6 @@ export default function Navigation({handleClick}) {
 
   return (
     <nav className="sui-StudioNav">
-      <input
-        className="sui-StudioNav-searchInput"
-        onChange={handleChange}
-        onFocus={handleFocus}
-        placeholder="Search"
-        type="search"
-        value={search}
-      />
-
       <ul className="sui-StudioNav-menu">
         {renderListFilteredBySearch({handleClick, search})}
       </ul>
@@ -82,5 +61,6 @@ export default function Navigation({handleClick}) {
 }
 
 Navigation.propTypes = {
+  search: PropTypes.string,
   handleClick: PropTypes.func
 }
