+++ packages/sui-studio/src/styles/_default.scss
@@ -1,4 +1,4 @@
 body {
   font-family: $ff-sans-serif;
-  font-size: $fz-body !important; // to force it even if infoJobs has a different rule in their theme
+  font-size: $fz-body;
 }
