+++ packages/sui-studio/src/styles/_settings.scss
@@ -81,6 +81,8 @@ $fgc-navbar--active: $c-white;
 $w-sidebar: 220px;
 $w-sidebar-collapsed: 48px;
 
+$h-search-input: 40px;
+
 $h-navHeader: 48px;
 $t-sidebar-and-main: 52px;
